Rename binaries.
