+++ cmd/webhook-manager/app/server.go
@@ -28,7 +28,7 @@ import (
 	"k8s.io/client-go/tools/clientcmd"
 	"k8s.io/klog"
 
-	"volcano.sh/volcano/cmd/admission/app/options"
+	"volcano.sh/volcano/cmd/webhook-manager/app/options"
 	"volcano.sh/volcano/pkg/admission/router"
 	"volcano.sh/volcano/pkg/version"
 )
