+++ installer/dockerfile/controller-manager/Dockerfile
@@ -13,10 +13,7 @@
 # limitations under the License.
 
 
-# The base image is created via `Dockerfile.base`, the base image is cached
-# since the required packages change very rarely.
-FROM volcanosh/vk-admission-base:latest
+FROM alpine:latest
 
-ADD vc-admission /vc-admission
-ADD gen-admission-secret.sh /gen-admission-secret.sh
-ENTRYPOINT ["/vc-admission"]
+ADD vc-controller-manager /vc-controller-manager
+ENTRYPOINT ["/vc-controller-manager"]
