+++ installer/dockerfile/webhook-manager/gen-admission-secret.sh
