+++ installer/helm/chart/volcano/values.yaml
@@ -1,8 +1,8 @@
 basic:
   image_tag_version: "latest"
-  controller_image_name: "volcanosh/vc-controllers"
+  controller_image_name: "volcanosh/vc-controller-manager"
   scheduler_image_name: "volcanosh/vc-scheduler"
-  admission_image_name: "volcanosh/vc-admission"
+  admission_image_name: "volcanosh/vc-webhook-manager"
   admission_secret_name: "volcano-admission-secret"
   scheduler_config_file: "config/volcano-scheduler.conf"
   image_pull_secret: ""
